v1.0.1

+++ services/tp_generate_device_service.go
@@ -28,16 +28,22 @@ func (*TpGenerateDeviceService) GetTpGenerateDeviceDetail(tp_generate_device_id 
 }
 
 // 获取列表
-func (*TpGenerateDeviceService) GetTpGenerateDeviceList(PaginationValidate valid.TpGenerateDevicePaginationValidate) (bool, []models.TpGenerateDevice, int64) {
-	var TpGenerateDevices []models.TpGenerateDevice
+func (*TpGenerateDeviceService) GetTpGenerateDeviceList(PaginationValidate valid.TpGenerateDevicePaginationValidate) (bool, []valid.TpGenerateDeviceListRsp, int64) {
+	var TpGenerateDevices []valid.TpGenerateDeviceListRsp
 	offset := (PaginationValidate.CurrentPage - 1) * PaginationValidate.PerPage
-	sqlWhere := "1=1"
-	if PaginationValidate.Id != "" {
-		sqlWhere += " and id = '" + PaginationValidate.Id + "'"
+	db := psql.Mydb.Model(&models.TpGenerateDevice{})
+	if PaginationValidate.AddFlag != "" {
+		db.Where("tp_generate_device.add_flag=?", PaginationValidate.AddFlag)
+	}
+	if PaginationValidate.DeviceCode != "" {
+		db.Where("tp_generate_device.device_code like ?", "%"+PaginationValidate.DeviceCode+"%")
 	}
 	var count int64
-	psql.Mydb.Model(&models.TpGenerateDevice{}).Where(sqlWhere).Count(&count)
-	result := psql.Mydb.Model(&models.TpGenerateDevice{}).Where(sqlWhere).Limit(PaginationValidate.PerPage).Offset(offset).Order("created_at desc").Find(&TpGenerateDevices)
+	db.Where("batch_id=?", PaginationValidate.BatchId).Count(&count)
+	result := db.Model(&models.TpGenerateDevice{}).
+		Select("tp_generate_device.id,tp_generate_device.add_flag,tp_generate_device.add_date,tp_generate_device.device_code,CASE WHEN ts_kv_latest.key='SYS_ONLINE' THEN '1' ELSE '0' END AS activate_flag").
+		Joins("left join ts_kv_latest on ts_kv_latest.entity_id = tp_generate_device.device_id  and ts_kv_latest.key='SYS_ONLINE' ").
+		Where("tp_generate_device.batch_id=?", PaginationValidate.BatchId).Limit(PaginationValidate.PerPage).Offset(offset).Find(&TpGenerateDevices).Order("tp_generate_device.created_time asc")
 	if result.Error != nil {
 		logs.Error(result.Error, gorm.ErrRecordNotFound)
 		return false, TpGenerateDevices, 0
@@ -57,6 +63,16 @@ func (*TpGenerateDeviceService) AddTpGenerateDevice(tp_generate_device models.Tp
 	return tp_generate_device, nil
 }
 
+//批量新增
+func (*TpGenerateDeviceService) AddBathTpGenerateDevice(tp_generate_devices []models.TpGenerateDevice) ([]models.TpGenerateDevice, error) {
+	result := psql.Mydb.Create(&tp_generate_devices)
+	if result.Error != nil {
+		logs.Error(result.Error, gorm.ErrRecordNotFound)
+		return tp_generate_devices, result.Error
+	}
+	return tp_generate_devices, nil
+}
+
 // 修改数据
 func (*TpGenerateDeviceService) EditTpGenerateDevice(tp_generate_device valid.TpGenerateDeviceValidate) bool {
 	result := psql.Mydb.Model(&models.TpGenerateDevice{}).Where("id = ?", tp_generate_device.Id).Updates(&tp_generate_device)
@@ -69,6 +85,11 @@ func (*TpGenerateDeviceService) EditTpGenerateDevice(tp_generate_device valid.Tp
 
 // 删除数据
 func (*TpGenerateDeviceService) DeleteTpGenerateDevice(tp_generate_device models.TpGenerateDevice) error {
+	var activatflag string
+	psql.Mydb.Select("add_flag").Where("id=?", tp_generate_device.Id).Find(&activatflag)
+	if activatflag == "1" {
+		return errors.New("已激活不能删除")
+	}
 	result := psql.Mydb.Delete(&tp_generate_device)
 	if result.Error != nil {
 		logs.Error(result.Error, gorm.ErrRecordNotFound)
@@ -80,7 +101,7 @@ func (*TpGenerateDeviceService) DeleteTpGenerateDevice(tp_generate_device models
 //生成设备表-批次表-产品表关联查询
 func (*TpGenerateDeviceService) generateDeviceProductBatch(id string) (map[string]interface{}, error) {
 	var gpb map[string]interface{}
-	sql := `select tgd.device_id as device_id,tgd.activate_flag as activate_flag ,tgd.token as token ,tgd.password as password ,tp.protocol_type as protocol_type ,tp.plugin as plugin, 
+	sql := `select tgd.device_id as device_id,tb.product_id as product_id,tgd.add_flag as add_flag ,tgd.token as token ,tgd.password as password ,tp.protocol_type as protocol_type ,tp.plugin as plugin, 
 	tb.access_address as access_address ,tp.serial_number as serial_number,tb.batch_number as serial_number,tp.device_model_id as device_model_id
 	from tp_generate_device tgd left join tp_batch tb on tgd.batch_id = tb.id left join tp_product tp on  tb.product_id = tp.id where tgd.id = ?`
 	result := psql.Mydb.Raw(sql, id).Scan(&gpb)
@@ -100,8 +121,8 @@ func (*TpGenerateDeviceService) ActivateDevice(generate_device_id string, asset_
 		logs.Error(err.Error())
 		return err
 	}
-	if gpb["activate_flag"] == "1" {
-		return errors.New("设备已激活，不能再次激活！")
+	if gpb["add_flag"] == "1" {
+		return errors.New("设备已添加，不能再次添加！")
 	}
 	var password = ""
 	if gpb["password"] != nil {
@@ -118,6 +139,7 @@ func (*TpGenerateDeviceService) ActivateDevice(generate_device_id string, asset_
 		Name:           device_name,
 		Protocol:       gpb["protocol_type"].(string),
 		Type:           gpb["device_model_id"].(string),
+		ProductId:      gpb["product_id"].(string),
 		DeviceType:     "1",
 		ProtocolConfig: "{}",
 		ChartOption:    "{}",
@@ -128,10 +150,14 @@ func (*TpGenerateDeviceService) ActivateDevice(generate_device_id string, asset_
 	if rsp_err != nil {
 		return rsp_err
 	}
+	// 获取当前日期并格式化为2006-01-02 15:04:05
+	t := time.Now().Format("2006-01-02 15:04:05")
+
 	//更新激活
 	var tp_generate_device = valid.TpGenerateDeviceValidate{
-		Id:           generate_device_id,
-		ActivateFlag: "1",
+		Id:      generate_device_id,
+		AddDate: t,
+		AddFlag: "1",
 	}
 	TpGenerateDeviceService.EditTpGenerateDevice(tp_generate_device)
 	return nil
