+++ services/tp_ota_service.go
@@ -4,6 +4,7 @@ import (
 	"ThingsPanel-Go/initialize/psql"
 	"ThingsPanel-Go/models"
 	valid "ThingsPanel-Go/validate"
+	"errors"
 
 	"github.com/beego/beego/v2/core/logs"
 	"gorm.io/gorm"
@@ -18,39 +19,76 @@ type TpOtaService struct {
 	TimeField []string
 }
 
-func (*TpOtaService) GetTpOtaList(PaginationValidate valid.TpOtaPaginationValidate) (bool, []models.TpOta, int64) {
-	var TpOtas []models.TpOta
-	offset := (PaginationValidate.CurrentPage - 1) * PaginationValidate.PerPage
-	db := psql.Mydb.Model(&models.TpOta{})
-	if PaginationValidate.ProductId != "" {
-		db.Where("product_id = ?", PaginationValidate.ProductId)
+//获取列表
+func (*TpOtaService) GetTpOtaList(PaginationValidate valid.TpOtaPaginationValidate) (bool, []map[string]interface{}, int64) {
+	sqlWhere := `select o.*,p.name as product_name from tp_ota o left join tp_product p on o.product_id=p.id where 1=1 `
+	sqlWhereCount := `select count(1) from tp_ota o left join tp_product p on o.product_id=p.id where 1=1`
+	var values []interface{}
+	var where = ""
+	if PaginationValidate.PackageName != "" {
+		values = append(values, "%"+PaginationValidate.PackageName+"%")
+		where += " and o.package_name like ?"
 	}
 	if PaginationValidate.Id != "" {
-		db.Where("id = ?", PaginationValidate.Id)
+		values = append(values, PaginationValidate.Id)
+		where += " and o.id = ?"
 	}
-	if PaginationValidate.PackageName != "" {
-		db.Where("package_name like ?", "%"+PaginationValidate.PackageName+"%")
+
+	if PaginationValidate.ProductId != "" {
+		values = append(values, PaginationValidate.ProductId)
+		where += " and o.product_id = ?"
 	}
+	sqlWhere += where
+	sqlWhereCount += where
 	var count int64
-	db.Count(&count)
-	result := db.Limit(PaginationValidate.PerPage).Offset(offset).Order("created_at desc").Find(&TpOtas)
+	result := psql.Mydb.Raw(sqlWhereCount, values...).Count(&count)
+	if result.Error != nil {
+		errors.Is(result.Error, gorm.ErrRecordNotFound)
+	}
+	var offset int = (PaginationValidate.CurrentPage - 1) * PaginationValidate.PerPage
+	var limit int = PaginationValidate.PerPage
+	sqlWhere += " offset ? limit ?"
+	values = append(values, offset, limit)
+	var otaList []map[string]interface{}
+	dataResult := psql.Mydb.Raw(sqlWhere, values...).Scan(&otaList)
+	if dataResult.Error != nil {
+		errors.Is(dataResult.Error, gorm.ErrRecordNotFound)
+	}
+	return true, otaList, count
+}
+
+//根据id获取升级包信息
+func (*TpOtaService) GetTpOtaVersionById(otaid string) (bool, models.TpOta) {
+	var TpOtas models.TpOta
+	result := psql.Mydb.Model(&models.TpOta{}).Where("id=?", otaid).Find(&TpOtas)
 	if result.Error != nil {
 		logs.Error(result.Error, gorm.ErrRecordNotFound)
-		return false, TpOtas, 0
+		return false, TpOtas
 	}
-	return true, TpOtas, count
+	return true, TpOtas
 }
 
 // 新增数据
-func (*TpOtaService) AddTpOta(tp_ota models.TpOta) (models.TpOta, error) {
+func (*TpOtaService) AddTpOta(tp_ota models.TpOta) (map[string]interface{}, error) {
+	var data map[string]interface{}
 	result := psql.Mydb.Create(&tp_ota)
 	if result.Error != nil {
 		logs.Error(result.Error, gorm.ErrRecordNotFound)
-		return tp_ota, result.Error
+		return data, result.Error
 	}
-	return tp_ota, nil
+	if err := psql.Mydb.Raw(`select o.*,p.name as product_name from tp_ota o left join tp_product p on o.product_id=p.id where o.id = ?`, tp_ota.Id).Scan(&data).Error; err != nil {
+		return data, err
+	}
+	return data, nil
 }
 func (*TpOtaService) DeleteTpOta(tp_ota models.TpOta) error {
+	var count int64
+	if err := psql.Mydb.Model(&models.TpOtaTask{}).Where("ota_id = ?", tp_ota.Id).Count(&count).Error; err != nil {
+		return err
+	}
+	if count != 0 {
+		return errors.New("存在升级任务不能删除固件")
+	}
 	result := psql.Mydb.Delete(&tp_ota)
 	if result.Error != nil {
 		logs.Error(result.Error)
